improve logging, add output/freeze/serialize
