+++ tests/test.py
@@ -8,7 +8,6 @@ import re
 import unittest
 
 import storable
-import storable.output
 
 P_ID = re.compile(r'[^a-zA-Z0-9]')
 
@@ -86,11 +85,11 @@ def make_function(deserializer, infile, outfile):
             'Deserialisation of %r did not equal the data '
             'given in %r' % (infile, outfile))
         try:
-            serialized_data = storable.output.serialize(data)
+            serialized_data = storable.freeze(data)
             reserialized_data = storable.thaw(serialized_data)
             if isinstance(data, dict):
-                storable.thaw(storable.output.modify_hash(serialized_data, 'fooobar_keyy', 'xy'))
-                storable.thaw(storable.output.modify_hash(serialized_data, 'abc', 45.12))
+                storable.thaw(storable.modify_hash(serialized_data, 'fooobar_keyy', 'xy'))
+                storable.thaw(storable.modify_hash(serialized_data, 'abc', 45.12))
         except RuntimeError as err:
             if 'recursion' in err.args[0].lower()\
                or repr(err).startswith('Recursion'):  # for python3.5+
